refactor: VoteDetailResponse에서 가중치 정보 제거
