+++ src/main/java/hanium/modic/backend/domain/vote/dto/VoteRewardResult.java
@@ -1,4 +1,4 @@
-package hanium.modic.backend.domain.vote.service;
+package hanium.modic.backend.domain.vote.dto;
 
 public record VoteRewardResult(boolean isCorrectAnswer, int currentStreak, boolean receivedTicket) {
 }
