+++ src/main/java/hanium/modic/backend/domain/vote/service/VoteRewardService.java
@@ -7,6 +7,7 @@ import hanium.modic.backend.common.error.exception.AppException;
 import hanium.modic.backend.common.property.property.VoteProperties;
 import hanium.modic.backend.domain.user.service.UserVoteStreakService;
 import hanium.modic.backend.domain.ticket.service.TicketService;
+import hanium.modic.backend.domain.vote.dto.VoteRewardResult;
 import hanium.modic.backend.domain.vote.entity.SimilarityVoteSummaryEntity;
 import hanium.modic.backend.domain.vote.enums.VoteDecision;
 import hanium.modic.backend.domain.vote.repository.SimilarityVoteSummaryRepository;
