+++ src/main/java/hanium/modic/backend/web/vote/dto/response/VoteDetailResponse.java
@@ -14,15 +14,6 @@ public record VoteDetailResponse(
 	@Schema(description = "생성된 이미지 URL (B)", example = "https://cloudfront.example.com/derived-image.jpg")
 	String derivedImageUrl,
 
-	@Schema(description = "찬성 가중치", example = "25")
-	Long approveWeight,
-
-	@Schema(description = "반대 가중치", example = "15")
-	Long denyWeight,
-
-	@Schema(description = "총 가중치", example = "40")
-	Long totalWeight,
-
 	@Schema(description = "투표 상태")
 	VoteStatus status
 ) {
